# Faster value binding.