Improve performance of value binding by checking for Valuers in With instead of Log.

- Always copy keyvals in BindValues.
- Export ContainsValuer and use in With.
- Add hasValuer bool field to withLogger.