@ChrisHines 2015-04-26
Comparing performance to the base branch I get:

```
benchmark                           old ns/op     new ns/op     delta
BenchmarkJSONLoggerSimple           7155          6870          -3.98%
BenchmarkJSONLoggerContextual       10005         9875          -1.30%
BenchmarkDiscard                    380           381           +0.26%
BenchmarkOneWith                    737           726           -1.49%
BenchmarkTwoWith                    940           936           -0.43%
BenchmarkTenWith                    2434          2360          -3.04%
BenchmarkPrefixLoggerSimple         2074          2016          -2.80%
BenchmarkPrefixLoggerContextual     3980          3950          -0.75%
BenchmarkValueBindingTimestamp      1181          1165          -1.35%
BenchmarkValueBindingCaller         2934          2856          -2.66%

benchmark                           old allocs     new allocs     delta
BenchmarkJSONLoggerSimple           19             19             +0.00%
BenchmarkJSONLoggerContextual       28             28             +0.00%
BenchmarkDiscard                    3              3              +0.00%
BenchmarkOneWith                    4              4              +0.00%
BenchmarkTwoWith                    4              4              +0.00%
BenchmarkTenWith                    4              4              +0.00%
BenchmarkPrefixLoggerSimple         5              5              +0.00%
BenchmarkPrefixLoggerContextual     11             11             +0.00%
BenchmarkValueBindingTimestamp      6              6              +0.00%
BenchmarkValueBindingCaller         6              6              +0.00%

benchmark                           old bytes     new bytes     delta
BenchmarkJSONLoggerSimple           812           812           +0.00%
BenchmarkJSONLoggerContextual       1159          1159          +0.00%
BenchmarkDiscard                    64            64            +0.00%
BenchmarkOneWith                    128           128           +0.00%
BenchmarkTwoWith                    192           192           +0.00%
BenchmarkTenWith                    704           704           +0.00%
BenchmarkPrefixLoggerSimple         145           145           +0.00%
BenchmarkPrefixLoggerContextual     402           402           +0.00%
BenchmarkValueBindingTimestamp      192           192           +0.00%
BenchmarkValueBindingCaller         176           176           +0.00%
```