@tsenart 2015-04-26
Hm, I think we might be going too far with premature optimisation. Not sure it's worth the extra complexity at this point.