@ChrisHines 2015-04-26
Haha, I pushed to soon. The newest version gets rid of the extra slice copy even when Valuers are present.

```
benchmark                           old ns/op     new ns/op     delta
BenchmarkJSONLoggerSimple           7155          6920          -3.28%
BenchmarkJSONLoggerContextual       10005         9835          -1.70%
BenchmarkDiscard                    380           369           -2.89%
BenchmarkOneWith                    737           723           -1.90%
BenchmarkTwoWith                    940           918           -2.34%
BenchmarkTenWith                    2434          2350          -3.45%
BenchmarkPrefixLoggerSimple         2074          2027          -2.27%
BenchmarkPrefixLoggerContextual     3980          3920          -1.51%
BenchmarkValueBindingTimestamp      1181          998           -15.50%
BenchmarkValueBindingCaller         2934          2590          -11.72%

benchmark                           old allocs     new allocs     delta
BenchmarkJSONLoggerSimple           19             19             +0.00%
BenchmarkJSONLoggerContextual       28             28             +0.00%
BenchmarkDiscard                    3              3              +0.00%
BenchmarkOneWith                    4              4              +0.00%
BenchmarkTwoWith                    4              4              +0.00%
BenchmarkTenWith                    4              4              +0.00%
BenchmarkPrefixLoggerSimple         5              5              +0.00%
BenchmarkPrefixLoggerContextual     11             11             +0.00%
BenchmarkValueBindingTimestamp      6              5              -16.67%
BenchmarkValueBindingCaller         6              5              -16.67%

benchmark                           old bytes     new bytes     delta
BenchmarkJSONLoggerSimple           812           812           +0.00%
BenchmarkJSONLoggerContextual       1159          1159          +0.00%
BenchmarkDiscard                    64            64            +0.00%
BenchmarkOneWith                    128           128           +0.00%
BenchmarkTwoWith                    192           192           +0.00%
BenchmarkTenWith                    704           704           +0.00%
BenchmarkPrefixLoggerSimple         145           145           +0.00%
BenchmarkPrefixLoggerContextual     402           402           +0.00%
BenchmarkValueBindingTimestamp      192           160           -16.67%
BenchmarkValueBindingCaller         176           144           -18.18%
```