@tsenart 2015-04-26
Hmmm, do we really need to make `BindValues` and `ContainsValuer` public?