@ChrisHines 2015-04-26
Not necessarily, but they could be useful to someone writing a custom Wither that wants to support Valuers.