@tsenart 2015-04-26
Makes me think if Wither should be public too.

On Sunday, April 26, 2015, Chris Hines <notifications@github.com> wrote:

> Not necessarily, but they could be useful to someone writing a custom
> Wither that wants to support Valuers.
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/ChrisHines/gokit/pull/3#issuecomment-96420201>.
>


-- 
Sent from Gmail Mobile