@ChrisHines 2015-04-26
Now that we've gotten this far into the API design and log.With has matured I think that is a valid question. I am on board with removing all of those pieces from the public API until we come across a use case that needs them exported.