@tsenart 2015-04-26
+1

On Sunday, April 26, 2015, Chris Hines <notifications@github.com> wrote:

> Now that we've gotten this far into the API design and log.With has
> matured I think that is a valid question. I am on board with removing all
> of those pieces from the public API until we come across a use case that
> needs them exported.
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/ChrisHines/gokit/pull/3#issuecomment-96421627>.
>


-- 
Sent from Gmail Mobile