@ChrisHines 2015-04-26
OK. I will merge this optimization and unexport those pieces when I get back to my PC.